tests/wpt/tests/css/css-shapes/shape-functions/path-function-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing the path() function</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-path">
<meta name="assert" content="Tests parsing of the path() function">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("clip-path", "path(evenodd, 'not-a-path')");
test_invalid_value("clip-path", "path(badfillrule, 'M10,10')");
</script>
</body>
</html>
